LayoutTests/platform/ios-sim-deprecated/http/tests/security/dataURL/xss-DENIED-to-data-url-sub-frame-2-level-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.

This tests that the main frame doesn't have access to a data: URL loaded in an iframe inside another data: URL loaded iframe.

Test 
PASS: Cross frame access to a data: URL 2 levels deep was denied.


--------
Frame: 'aFrame'
--------
Inner iframe.



--------
Frame: '<!--framePath //aFrame/<!--frame0-->-->'
--------
PASS: Cross frame access to a data: URL 2 levels deep was denied.

Inner-inner iframe.
